Admin LogIn and Admin Restriction

+++ pages/admin/admin.php
@@ -1,13 +1,19 @@
 <?php
+    session_start();
+    if (isset($_SESSION['sessAdminId'])) {
 
-    if(isset($_GET['aId'])){
-        $admins = $admin->delete('adminId',$_GET['aId']);	
+        if(isset($_GET['aId'])){
+            $admins = $admin->delete('adminId',$_GET['aId']);	
+        }
+
+        $admins = $admin->findAll();
+
+        $title = "MeKart";
+
+        $content = loadTemplate("templates/adminAdmin.php", ['admins' => $admins]);
     }
 
-    $admins = $admin->findAll();
-
-    $title = "MeKart";
-
-    $content = loadTemplate("templates/adminAdmin.php", ['admins' => $admins]);
+    else
+        include 'adminLogIn.php';
 
 ?>
