+++ pages/admin/product.php
@@ -1,12 +1,17 @@
 <?php
-    $title = "MeKart";
+    session_start();
+    if (isset($_SESSION['sessAdminId'])) {
+        $title = "MeKart";
 
-    if(isset($_GET['pId'])){
-        $products = $product->delete('productId',$_GET['pId']);	
+        if(isset($_GET['pId'])){
+            $products = $product->delete('productId',$_GET['pId']);	
+        }
+
+        $products = $product->joinTable('vendor','vendorId', 'vendorId');
+
+        $content = loadTemplate("templates/adminProduct.php", ['products' => $products]);
     }
-
-    $products = $product->joinTable('vendor','vendorId', 'vendorId');
-
-    $content = loadTemplate("templates/adminProduct.php", ['products' => $products]);
+    else
+        include 'adminLogIn.php';
 
 ?>
